// frontend/generated/com/example/app/endpoint/Options.ts
import Rate from './Rate';

/**
 * This module is generated from com.example.app.endpoint.Options.
 * All changes to this file are overridden. Please consider to make changes in the corresponding Java file if necessary.
 * @see {@link file:///Users/artur/workspace/labs/mortgage-calculator/src/main/java/com/example/app/endpoint/Options.java}
 */
export default interface Options {
  amount: number;
  paybackTimeMonths: number;
  rate: Rate;
}