// frontend/generated/com/example/app/endpoint/Rate.ts
/**
 * This module is generated from com.example.app.endpoint.Rate.
 * All changes to this file are overridden. Please consider to make changes in the corresponding Java file if necessary.
 * @see {@link file:///Users/artur/workspace/labs/mortgage-calculator/src/main/java/com/example/app/endpoint/Rate.java}
 */
export default interface Rate {
  defaultRate: boolean;
  margin: number;
  name: string;
  rate: number;
}